Fix local alias handling in type emit.
Previously, tsickle would resolve all aliased symbols when printing types. The
idea was that we'd resolve all symbols to their final alias in all places,
which then would hopefully be canonical, and handle aliasing based on that.

This however meant that we would often use the final symbol of a chain of
aliases, instead of the immediately imported symbol and module name. That is
problematic in case the final name is synthetic and doesn't exist in Closure
Compiler - e.g. when it comes from a `.d.ts` file that does "interesting"
things to present the right set of types to TypeScript.

tsickle also didn't handle a special case of alias: type aliases (as opposed to
symbol aliases).

This change fixes both those situations:
* instead of eagerly resolving aliases, it creates aliases for all local
  imported symbols.
* it special cases type aliases.

PiperOrigin-RevId: 430652229